Lab tidy
Most of these labs are out of date and not supported. They've now been deleted with the exception of some labs that were recently updated.

Rather than delete the labs entirely, the readmes have been updated to point to an equivalent module or learning path on Microsoft Learn. This is because the Azure Portal still points to these labs, but we are awaiting an update to point to learn instead.